Fix migration schema drift, redesign PDF documents, polish storefront header
Migrations
- Add missing migrations for tables that existed live but were never
  captured (orders, delivery_addresses, otp_codes, category_products),
  and for columns added directly to the DB (pcs/price_carton on
  products, status on stores) — a fresh `migrate` on an empty database
  now runs end to end without errors, verified by migrating a scratch
  database from zero
- Fix several existing migrations that referenced columns/tables that
  no longer matched the live schema, so the migration history is now a
  coherent, shareable source of truth for the team

PDF documents (Dompdf)
- Rebuild the invoice ("Facture") template from scratch — it depended
  on a stylesheet (assets/css/invoice-details.css) that had been
  deleted during the earlier rebrand cleanup, so every custom class
  had zero styling; replaced with a self-contained, on-brand design
- Redesign the daily report and movements report to match the SMH
  terracotta theme instead of a leftover blue palette, and fix a real
  Dompdf incompatibility: it doesn't reliably support CSS
  `linear-gradient()` backgrounds or `display: flex`, which was making
  the summary cards render invisible (white text on a background that
  silently failed to render) and causing large layout gaps
- Remove Google Fonts `<link>` usage from the daily report — Dompdf
  renders server-side and tries to fetch/cache web fonts locally,
  which was crashing the export outright; switched to safe system fonts

Storefront header
- Add a product search bar to the shared header, wired to the
  catalogue's existing client-side filter via a `?search=` query param
- Rework the header layout into a balanced single-line design (logo /
  centered menu / search + account icons), fixing a real overflow bug
  right at the `lg` breakpoint where the menu, search bar, and icons
  didn't fit together

+++ database/migrations/2024_08_09_073416_create_sales_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->foreign('product_id')->references('id')->on('products')->onUpdate('cascade')->onDelete('restrict');
             $table->foreignId('store_id')->constrained('stores')->restrictOnDelete();
             $table->integer('quantity');
-            $table->float('prix', 20, 2);
-            $table->float('prixTotal', 20, 2);
-            $table->float('interet', 20, 2);
+            $table->double('prix', 20, 2);
+            $table->double('prixTotal', 20, 2);
+            $table->double('interet', 20, 2);
             $table->timestamps();
         });
     }
